% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{Abstract}
\alias{Abstract}
\title{Abstract:
  uri: <http://salt.semanticauthoring.org/ontologies/sro#Abstract>
comment: an abstract of a science article}
\format{An object of class \code{identifier} of length 4.}
\usage{
Abstract
}
\description{
Abstract:
  uri: <http://salt.semanticauthoring.org/ontologies/sro#Abstract>
comment: an abstract of a science article
}
\keyword{datasets}
