% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-access-low-level.R
\name{add_data}
\alias{add_data}
\title{Add data to a repository}
\usage{
add_data(server_access_options, repository, data,
  data_format = "application/x-trig", update = FALSE)
}
\arguments{
\item{server_access_options}{the access details to an RDF4J store, created by
\code{server_access_options}}

\item{repository}{the name of the repository, where to add data to}

\item{data}{the RDF data containing the triples to be submitted}

\item{data_format}{data format, one of {"application/x-trig"}}
}
\description{
Add data to a repository
}
