% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uri.R
\name{add_prefix}
\alias{add_prefix}
\title{Adds the openBiodiv prefix to a string to create a proper URI}
\usage{
add_prefix(string)
}
\arguments{
\item{string}{(character) to be pasted to}
}
\value{
URI
}
\description{
Adds the openBiodiv prefix to a string to create a proper URI
}
