% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{Address}
\alias{Address}
\title{To specify the components of the delivery address for the object}
\format{An object of class \code{identifier} of length 4.}
\usage{
Address
}
\description{
To specify the components of the delivery address for the object
}
\keyword{datasets}
