% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{alt_label}
\alias{alt_label}
\title{SKOS Alternative Label}
\format{An object of class \code{identifier} of length 4.}
\usage{
alt_label
}
\description{
SKOS Alternative Label
}
\keyword{datasets}
