% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\docType{data}
\name{AnonRDF}
\alias{AnonRDF}
\title{Create a list of RDF statements that all share the same blank subject node}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
AnonRDF
}
\description{
Create a list of RDF statements that all share the same blank subject node
}
\seealso{
Other anonymous rdf: \code{\link{is.AnonRDF}}
}
\concept{anonymous rdf}
\keyword{datasets}
