% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{article_dumper}
\alias{article_dumper}
\title{Article Dumper}
\usage{
article_dumper(journal = "BDJ", fromdate = "01/01/2010")
}
\arguments{
\item{journal}{the alias of the journal that is to be dumped.
allowable are "BDJ", "ZooKeys", "PhytoKeys". if missing dumps BDJ}

\item{fromdate}{from which date on. If missing starts from 1.jan.2010}
}
\value{
vector of new files that have been dumped
}
\description{
Article Dumper
}
