% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{article_publication_date.filename}
\alias{article_publication_date.filename}
\title{What is the publication date of a scholarly article?}
\usage{
article_publication_date.filename(filename, ...)
}
\arguments{
\item{filename}{filename with XML2 document}

\item{...}{additional parameter to pass to the extractor (like XML schema)}
}
\value{
publication date of the article as a string
}
\description{
What is the publication date of a scholarly article?
}
