% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{article_publication_date}
\alias{article_publication_date}
\title{What is the publication date of a scholarly article?}
\usage{
article_publication_date(xml_document,
  day_path = "//article-meta/pub-date/day",
  month_path = "//article-meta/pub-date/month",
  year_path = "//article-meta/pub-date/year")
}
\arguments{
\item{xml_document}{XML2 document containing the article}

\item{day_path}{the XPATH expression containing the day of publication
TODO : pass these as some sort of schema's (for refactoring)}
}
\value{
date of publication as string
}
\description{
What is the publication date of a scholarly article?
}
