% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plazi.R
\name{as.data.frame.plazi_feed}
\alias{as.data.frame.plazi_feed}
\title{Plazi Feed to Dataframe Converter}
\usage{
\method{as.data.frame}{plazi_feed}(feed)
}
\arguments{
\item{feed}{plazi_feed, xml_document, the Plazi feed.}
}
\value{
a `data.frame` and a `plazi_infotable` object containing tabular
  information about Plazi treatments
}
\description{
Converts the Plazi feed XML to a data frame with interesting columns.
}
\examples{
a_plazi_feed = plazi_feed()
xml2::write_xml(a_plazi_feed, file = "~/tmp/plazi_feed.xml")
b_plazi_feed = plazi_feed("/home/viktor2/tmp/plazi_feed.xml")
plazi_info = as.data.frame(b_plazi_feed)

}
