% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdfization.R
\name{as.rdf}
\alias{as.rdf}
\title{Convert an Entity to RDF}
\usage{
as.rdf(x)
}
\arguments{
\item{x}{Converts the atoms to appropriate triples. and connects to parent ID via
po:contains.}

\item{x}{an entity to convert to RDF}
}
\value{
RDF
}
\description{
Convert an Entity to RDF
}
