% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxonomicName.R
\name{as.TaxonomicName}
\alias{as.TaxonomicName}
\title{Transform an Object to TaxonomicName}
\usage{
as.TaxonomicName(x, ...)
}
\arguments{
\item{x}{the object to be transformed to a taxonomic name}

\item{...}{additional parameters passed to the appropriate consturctor}
}
\value{
TaxonomicName
}
\description{
Generic function.
}
