% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxonomicNameUsage.R
\name{as.TaxonomicName.TaxonomicNameUsage}
\alias{as.TaxonomicName.TaxonomicNameUsage}
\title{Transform a TaxonomicNameUsage to a TaxonomicName}
\usage{
\method{as.TaxonomicName}{TaxonomicNameUsage}(TNU)
}
\arguments{
\item{TNU}{TaxonomicNameUsage}
}
\value{
TaxonomicName of the appropriate sub-class
}
\description{
Transform a TaxonomicNameUsage to a TaxonomicName
}
