% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdfization-author.R
\name{author_rule.1}
\alias{author_rule.1}
\title{Author Matches According to Author Rule 1}
\usage{
author_rule.1(collab, surname, given_name, email, reference, affiliation, label)
}
\arguments{
\item{collab, }{surname, given_name, email, reference, affiliation}
}
\value{
a data.frame with person matches of class SearchResult
}
\description{
Looks up an author's ID by his or her email.
}
\details{
Idea: first letter of first name AND
      exact match of email AND
      fuzzy match of last name
}
