% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdfization-author.R
\name{author_rule.2}
\alias{author_rule.2}
\title{Person Matches According to Author Rule 2}
\usage{
author_rule.2(collab, surname, given_name, email, reference, affiliation, label)
}
\arguments{
\item{collab, }{surname, given_name, email, reference, affiliation}
}
\value{
a data.frame with person matches of class SearchResult
}
\description{
Match against existing affiliations.
Note: one author may have multiple affiliations.
}
\details{
Idea: first letter of first name AND
      fuzzy-and-match of last name AND
      fuzzy-and-match of affiliation AND
}
