% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdfization-author.R
\name{author_rule.3}
\alias{author_rule.3}
\title{Person Matches According to Author Rule 3}
\usage{
author_rule.3(collab, surname, given_name, email, reference, affiliation, label)
}
\arguments{
\item{affiliation}{}
}
\value{
a data.frame with person matches of class SearchResult
}
\description{
For collaborative authors, match the label
}
