% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdfization-author.R
\name{author_rule.person}
\alias{author_rule.person}
\title{Execute all author rules on a Person object}
\usage{
author_rule.person(person)
}
\arguments{
\item{person}{an object of class Person}
}
\value{
search results data.frame as returned by the rules
}
\description{
Useful to get all URI's that correspond to a Person.
}
\details{
The function works by creating all possible combinations of the Person's
fields and running them against all the rules.
}
