% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nactem.R
\name{binomialize}
\alias{binomialize}
\title{Binomialize Taxonomic Name}
\usage{
binomialize(taxname)
}
\arguments{
\item{taxname}{character vector with name parts with name to binomialize}
}
\value{
just the binomial name
}
\description{
Binomialize Taxonomic Name
}
\examples{

binomialize("Agapostemon (Agapostemon) sericeus (Förster, 1771)")
}
