% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdfization-body.R
\name{body_extractor}
\alias{body_extractor}
\title{Body Extractor}
\usage{
body_extractor(body_comp, metadata)
}
\arguments{
\item{body_comp}{a component (list of an XML2 node and an id
vector) containing the body of the article}

\item{metadata}{a journal article object}
}
\value{
triples
}
\description{
Extracts knowledge from the body of an article
}
