% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosslinking_functions.R
\name{bold_genbank_serializer}
\alias{bold_genbank_serializer}
\title{Serialization of BOLD record ids/BINS/Genbank ids in different article sections (a separate function like institution_serializer)}
\usage{
bold_genbank_serializer(tt, atoms, identifiers, publisher_id, journal_id,
  plazi_doc, doi = doi, article_id = article_id)
}
\description{
Serialization of BOLD record ids/BINS/Genbank ids in different article sections (a separate function like institution_serializer)
}
