% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{boldsystems}
\alias{boldsystems}
\title{datacite boldsystems}
\format{An object of class \code{identifier} of length 4.}
\usage{
boldsystems
}
\description{
datacite boldsystems
}
\keyword{datasets}
