% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plazi.R
\name{check_download_status}
\alias{check_download_status}
\title{Check Download Status of Plazi Treatment}
\usage{
check_download_status(plazi_id, corpus_dir, xml_schema)
}
\arguments{
\item{plazi_id}{character vector, with Plazi ID's whose status ought to be checked}

\item{corpus_dir}{string, where the Plazi treatment files have been downloaded.}

\item{xml_schema}{an object of class `xml_schema`, containing information
about the xml_schema}
}
\value{
logical vector
}
\description{
Looks in a Plazi corpus directory and updates a Plazi information frame
with it. Tells you which treatments have been downloaded.
}
\details{
In description below, "string" refers to a character vector of length 1.
}
\examples{
download_status = check_download_status(plazi_info$id,
  corpus_dir = "/media/obkms/plazi-corpus-xml", xml_schema = taxonx)


}
