% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{collection_code_extractor}
\alias{collection_code_extractor}
\title{Extract Collection Code}
\usage{
collection_code_extractor(paper_id, cc_xml, article_id)
}
\arguments{
\item{paper_id}{URI of the `fabio:ResearchPaper` of the paper}

\item{cc_xml}{XML2 node of the collection code usage}
}
\value{
list of triples
}
\description{
Extract Collection Code
}
