% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{comp_func_stats}
\alias{comp_func_stats}
\title{Compute Statistics For a Particular Function}
\usage{
comp_func_stats(fname, Log)
}
\arguments{
\item{fname}{the name of the function for which you want to compute statistics}
}
\value{
a list of statistics
}
\description{
Computes:
  Total calls
  percentage of different statuses of the particular type
}
