% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{compute_statistics_from_log}
\alias{compute_statistics_from_log}
\title{Compute Various Iteration Statistics}
\usage{
compute_statistics_from_log(Log)
}
\arguments{
\item{Log}{a log data.frame to work with}
}
\value{
a \emph{list} of statistics
}
\description{
Compute Various Iteration Statistics
}
