% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{create_server_options}
\alias{create_server_options}
\title{Create an object with the settings needed to access the server}
\usage{
create_server_options(protocol, server_add, authentication = "basic_http",
  userpwd, api_key, secret, repository = "")
}
\arguments{
\item{protocol}{a string, either "http://" or "https://". Will be prepended to the endpoint URL.}

\item{server_add}{a string, the address part (without the protocol) of the RDF4J server. Please, do not use trailing slash!}

\item{authentication}{a string, either "basic_http" or "api".
In the case of "basic_http", you are further required to supply a username-password combination.
In the case of "api", you are further required to supply an API key and the associated secret.}

\item{userpwd}{a string, a user-password combination, used in case of basic HTTP authentication, e.g. "user:pass".}

\item{api_key}{a string, the API key used for API-style authentication.}

\item{secret}{a string, the secret string corresponding to the API key needed for API-style authentication.}

\item{repository}{a string, an optional parameter specifying the repository in the graph database that we will be accessing}
}
\description{
This function is used to create an \code{options} variable, containing all the configuration
needed to access the RDF4J endpoint: protocol, server address, authentication method, and
credentials.
}
\examples{
\dontrun{options = create_server_options(protocol = "http://", server_add = "213.191.204.69:7777/graphdb", authentication = "basic_http", userpwd = "deliberately_changed_username:fake_password"), repository = "OBKMS"}
\dontrun{options = create_server_options(protocol = "https://", server_add = "rdf.s4.ontotext.com/4937448214/OBKMS", authentication = "api", api_key = "wrong_api_key", secret = "no_secret"  )}
}

