% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{create_server_options_env}
\alias{create_server_options_env}
\title{Creates the server options from a YAML file and the environment}
\usage{
create_server_options_env(filename)
}
\description{
Creates the server options from a YAML file and the environment
}

