% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosslinking_functions.R
\name{crosslinker}
\alias{crosslinker}
\title{Crosslinker extracts external identifiers within the xml which enables crosslinking between OpenBiodiv and other systems
Can add more types of identifiers to be extracted}
\usage{
crosslinker(file)
}
\description{
Crosslinker extracts external identifiers within the xml which enables crosslinking between OpenBiodiv and other systems
Can add more types of identifiers to be extracted
}
