% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{dc_title}
\alias{dc_title}
\title{DC Elements Title}
\format{An object of class \code{identifier} of length 4.}
\usage{
dc_title
}
\description{
DC Elements Title
}
\keyword{datasets}
