% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{Distribution}
\alias{Distribution}
\title{Distribution}
\format{An object of class \code{identifier} of length 4.}
\usage{
Distribution
}
\description{
Distribution
}
\keyword{datasets}
