% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdfization-metadata.R
\name{document_components}
\alias{document_components}
\title{Break Down a XML object into its Components}
\usage{
document_components(xml, document_component_xpath)
}
\arguments{
\item{xml}{the xml resource}

\item{document_component_xpath}{datapase of xpaths where the document components
are found in the corresponding XML schema}
}
\value{
a list of XML nodes and ID's corresponding to each of the elements
  in the `document_component_xpath` vector
}
\description{
Takes a top-level XML object (e.g. a paper) and a vector of XPATH locations,
indicating where its subcomponents are.
This function returns a list of XML objects that are found at the XPATH
locations (that can later be processed by lower-level extractors) with their
IDs.
}
