% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentComponent.R
\name{DocumentComponent}
\alias{DocumentComponent}
\title{DocumentComponent Constructor}
\usage{
DocumentComponent(node, atom_location, component_class)
}
\arguments{
\item{node}{the XML node that will be converted to a `DocumentComponent`}

\item{atom_location}{a vector of XPATH where the objects will be sought after}

\item{component_class}{the class that will be assigned to the object}
}
\value{
an object of class \code{component_class}
}
\description{
This constructor creates a document component from an XML node.
}
\details{
TODO Note: a document component is made up of atoms. The atoms of a component
may be literals as well as URI's. Currently, only literals are implemented.
}
