% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{dwc_authorship}
\alias{dwc_authorship}
\title{Dwc Authorshiop}
\format{An object of class \code{identifier} of length 4.}
\usage{
dwc_authorship
}
\description{
Dwc Authorshiop
}
\keyword{datasets}
