% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{dwc_kingdom}
\alias{dwc_kingdom}
\title{Dwc Kingdom}
\format{An object of class \code{identifier} of length 4.}
\usage{
dwc_kingdom
}
\description{
Dwc Kingdom
}
\keyword{datasets}
