% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{dwc_subspecies_ep}
\alias{dwc_subspecies_ep}
\title{Dwc genus}
\format{An object of class \code{identifier} of length 4.}
\usage{
dwc_subspecies_ep

dwc_subspecies_ep
}
\description{
Dwc genus

Dwc genus
}
\keyword{datasets}
