% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{exact_match}
\alias{exact_match}
\title{SKOS exact match}
\format{An object of class \code{identifier} of length 4.}
\usage{
exact_match
}
\description{
SKOS exact match
}
\keyword{datasets}
