% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-access-high-level.R
\name{execute_openbiodiv_query}
\alias{execute_openbiodiv_query}
\title{Procedure That Executes a Query Against OBKMS}
\usage{
execute_openbiodiv_query(query, replacement = NA,
  result_format = "data.frame")
}
\arguments{
\item{query}{the unprefixed, unsubstituted SPARQL query (character vector)}

\item{replacement}{a named character vector of replacements}

\item{result_format}{how results should be returned. One of
c("data.frame")}
}
\value{
whatever the OpenBiodiv SPARQL endpoint returns, mediated by
  `result_format`
}
\description{
This procedure takes a query without prefixes, adds the prefixes, carries out
the replacements, and executes the query against the predefined OpenBiodiv
endpoint.
}
