% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{ExpressionCollection}
\alias{ExpressionCollection}
\title{an article in a science journal}
\format{An object of class \code{identifier} of length 4.}
\usage{
ExpressionCollection
}
\description{
an article in a science journal
}
\keyword{datasets}
