% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdfization-metadata.R
\name{extract_keywords}
\alias{extract_keywords}
\title{Extract Keywords from a Keyword Group}
\usage{
extract_keywords(XML, Identifier,
  XPath = yaml::yaml.load_file(obkms$config$keywords_db_xpath),
  keyword_scheme = obkms$parameters$Vocabularies$Subject_Classification_Terms)
}
\arguments{
\item{XML}{a list containing the XMl-node of the keyword-group.
The following fields need to be present: `node`, `document`}

\item{Identifier}{a list of identifiers used in the triple generation
The following fields need to be present: `paper`}

\item{XPath}{a list containing the locations of literals. Has default.}

\item{keyword_scheme}{the classification scheme in which the keyword is.
Has default.}
}
\value{
triples pertaining to the keyword
}
\description{
Takes care of the language as well.
}
