% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{Figure}
\alias{Figure}
\title{Figure:
 uri: <http://purl.org/spar/doco/Figure>
  comment: a figure in a science article}
\format{An object of class \code{identifier} of length 4.}
\usage{
Figure
}
\description{
Figure:
 uri: <http://purl.org/spar/doco/Figure>
  comment: a figure in a science article
}
\keyword{datasets}
