% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{find_atoms}
\alias{find_atoms}
\title{Simpler version of find literals}
\usage{
find_atoms(xml, xpath)
}
\arguments{
\item{xml}{the XML node}

\item{xpath}{the atom locations as a named character vector}
}
\value{
list
}
\description{
Finds the Atoms in a XML Node
}
