% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{find_literals}
\alias{find_literals}
\title{Finds the atoms in an xml node}
\usage{
find_literals(xml, x)
}
\arguments{
\item{xml}{the XML node}

\item{x}{xpath expressions}
}
\value{
a vector of atoms
}
\description{
Finds the atoms in an xml node
}
