% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{frbr_part}
\alias{frbr_part}
\title{e-Issn property}
\format{An object of class \code{identifier} of length 4.}
\usage{
frbr_part
}
\description{
e-Issn property
}
\keyword{datasets}
