% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdfization-front-matter.R
\name{front_matter_extractor}
\alias{front_matter_extractor}
\title{Front Matter Extractor}
\usage{
front_matter_extractor(front_matter_comp, title_comp, abstract_comp, metadata)
}
\arguments{
\item{front_matter_comp}{a component (list of an XML2 node and an id
vector) containing the front matter of the article}

\item{title_comp}{a component containing the title of the article}

\item{abstract_comp}{a component containing the abstract of the article}

\item{metadata}{a journal article object}
}
\value{
triples
}
\description{
Extracts the front matter of an article.
}
