% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{FrontMatter_extractor}
\alias{FrontMatter_extractor}
\title{Front Matter Extractor}
\usage{
FrontMatter_extractor(comp, article)
}
\arguments{
\item{comp}{A list of components containing the front matter.}

\item{article}{a journal article object}
}
\value{
triples
}
\description{
Extracts the front matter as RDF from a list of components
}
