% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{gbif_dataset}
\alias{gbif_dataset}
\title{datacite gbif dataset}
\format{An object of class \code{identifier} of length 4.}
\usage{
gbif_dataset
}
\description{
datacite gbif dataset
}
\keyword{datasets}
