% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosslinking_functions.R
\name{gbif_taxonomy_mapping}
\alias{gbif_taxonomy_mapping}
\title{Maps ScNames from xmls to ScNames from GBIF}
\usage{
gbif_taxonomy_mapping(scName, collection = checklistCol)
}
\description{
Maps ScNames from xmls to ScNames from GBIF
}
