% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{GenBankAccession}
\alias{GenBankAccession}
\title{GenBankAccession}
\format{An object of class \code{identifier} of length 4.}
\usage{
GenBankAccession
}
\description{
GenBankAccession
}
\keyword{datasets}
