% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plazi.R
\name{generate_aria2c_input}
\alias{generate_aria2c_input}
\title{Generate `aria2c` Input}
\usage{
generate_aria2c_input(id, corpus_dir)
}
\arguments{
\item{corpus_dir}{}
}
\value{
character vector, containing the `aria2c` input file
}
\description{
`aria2c` is a download tool. Its advantage over 'wget' is that it is fast
thanks to using multiple connections.
}
\details{
This function generates an input file for `aria2c` for Plazi treatments.

WILL DOWNLOAD THE ONES THAT ARE FALSE!
}
\examples{

}
