% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plazi.R
\name{generate_aria2c_input_images}
\alias{generate_aria2c_input_images}
\title{Generate Aria Input For Images}
\usage{
generate_aria2c_input_images(image_list)
}
\arguments{
\item{image_list}{named list of links to images}
}
\value{
character vector; `aria2c` input file
}
\description{
Generate Aria Input For Images
}
\examples{
generate_aria2c_input_images(plazi_images(treatments$id, directory = plazi_corpus_dir))

}
