% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{get_filename_extension}
\alias{get_filename_extension}
\title{Gets the filenmae extension}
\usage{
get_filename_extension(filename)
}
\arguments{
\item{filename}{}
}
\value{
extension
}
\description{
Gets the filenmae extension
}
\examples{
get_filename_extension("https://zenodo.org/record/1163869/files/figure.png")

}
