% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{get_label}
\alias{get_label}
\title{Gets the RDF Label (rdfs:label) of an Object}
\usage{
get_label(x, ...)
}
\arguments{
\item{x}{Object of which we need the label for.}
}
\value{
Character. The RDF label of the object
}
\description{
Gets the RDF Label (rdfs:label) of an Object
}
\seealso{
get_label.LatinName
}
