% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxonomicName.R
\name{get_label.TaxonomicName}
\alias{get_label.TaxonomicName}
\title{Gets the Label of a Latinized Name}
\usage{
\method{get_label}{TaxonomicName}(x)
}
\arguments{
\item{x}{Latininzed name (LatinName).}
}
\value{
Character. The label of the Latinized name.
}
\description{
Gets the Label of a Latinized Name
}
