% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxonomicNameUsage.R
\name{get_label.TaxonomicNameUsage}
\alias{get_label.TaxonomicNameUsage}
\title{Gets the Label of the Taxonomic Name Mentioned by the Taxonomic Name Usage}
\usage{
\method{get_label}{TaxonomicNameUsage}(x)
}
\arguments{
\item{x}{Taxonomic name usage.}
}
\value{
Character. The label of the mentioned name
}
\description{
Gets the Label of the Taxonomic Name Mentioned by the Taxonomic Name Usage
}
