% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxonomicName.R
\name{get_label_TaxonomicName}
\alias{get_label_TaxonomicName}
\title{Get the Label (Monimial, Bionomial, or Trinomial) from Taxonomic Name Atoms}
\usage{
get_label_TaxonomicName(kingdom, phylum, class, order, family, genus, subgenus,
  species, subspecies, authorship, secundum_literal)
}
\arguments{
\item{secundum_literal}{}
}
\value{
the correct way to use the name
}
\description{
Get the Label (Monimial, Bionomial, or Trinomial) from Taxonomic Name Atoms
}
