% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{get_or_set_obkms_id}
\alias{get_or_set_obkms_id}
\title{NOT USED ANYMORE
Get the OBKMS Id of an XML node, if not set, set it.}
\usage{
get_or_set_obkms_id(node)
}
\arguments{
\item{node}{the XML node from which to take the ID; cannot be missing!}
}
\value{
the local id (not an identifier object)
}
\description{
Does not do any database lookups.
}
