% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plazi.R
\name{get_plazi_treatment_language}
\alias{get_plazi_treatment_language}
\title{Get Plazi Treatment Language}
\usage{
get_plazi_treatment_language(id, directory)
}
\arguments{
\item{id}{character. treatment id's.}

\item{directory}{string. where the treatments are stored.}
}
\value{
character, languages. NA if file cannot be read.
}
\description{
Get Plazi Treatment Language
}
\examples{
a_PlaziFeed = PlaziFeed()
plazi_infotable = as.data.frame(a_PlaziFeed)
download_plazi_treatments(uri = plazi_infotable$taxonx[1:100], id = plazi_infotable$id[1:100], dir = "/home/viktor2/tmp", logdir = "/home/viktor2/tmp")
get_plazi_treatment_language(id = plazi_infotable$id[1:100], directory = "/home/viktor2/tmp")

}
