% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-access-low-level.R
\name{get_protocol_version}
\alias{get_protocol_version}
\title{REPOSITORY INFORMATION
Test connectivity to the graph database and get the communication protocol
version.}
\usage{
get_protocol_version(options)
}
\arguments{
\item{options}{a list, containing the graph database connectivity options,
returned by the helper function \code{server_access_options}.}
}
\value{
a numeric, containing the protocol version if connectivity is OK.
}
\description{
This function tests the connectivity to the graph database. If there is
connectivity it will return the protocol version as a numeric.
If there is no connectivity, an error condition will be set.
}
\examples{

\dontrun{ get_protocol_version( options ) }

}
