% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-access-low-level.R
\name{GET_query}
\alias{GET_query}
\title{Use the GET protocol to execute a graph database query}
\usage{
GET_query(options, repo_id, query, results_format = "CSV",
  query_ln = "sparql", infer = TRUE, varbindings = "", timeout = "")
}
\arguments{
\item{options}{a list returned by \code{server_access_options}.}

\item{repo_id}{a string holding the repository id as returned by \code{get_repositories}.}

\item{results_format}{a string, indicating the type of results required. Currently supported: "XML", "CSV".
If "XML" is specified, the return value is an XML string, ready to be parsed as a document.
If "CSV" is specified, the return variable is already parsed as a data.frame.}

\item{infer}{a boolean, indicating whether we want to use inference.}

\item{varbindings}{currently unused}

\item{timeout}{currently unused}

\item{query_language}{a string: "SPARQL" or "SERQL" (untested)}
}
\description{
This function submits a query to a previously configured graph database
(configuration is speciefied in the \code{options} parameter) using the GET
protocol. The query string (in SPARQL - tested, or SERQL - not yet tested) as
well as the return format of the results are passed as parameters.
}
\note{
This function has currently only been tested with the S4 semantic suite of
  Ontotext, as there seems to be some bug with the standalone GraphDB
  installation (in GraphDB, not in the package!).
}
\examples{
\dontrun{r = GET_query( options, "plazi", query, "CSV" )}
\dontrun{r = GET_query( options, "plazi", query, "XML" )}
}
