% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biology.R
\name{get_scientific_name_or_tcl}
\alias{get_scientific_name_or_tcl}
\title{Get the Label (Monimial, Bionomial, or Trinomial Name, or Taxonomic Concept Label) from Taxonomic Name Atoms}
\usage{
get_scientific_name_or_tcl(kingdom = NA, phylum = NA, class = NA,
  order = NA, family = NA, subfamily = NA, genus = NA,
  subgenus = NA, species = NA, subspecies = NA, authorship = NA,
  secundum_literal = NA)
}
\arguments{
\item{secundum_literal}{}
}
\value{
the correct way to use the name
}
\description{
Get the Label (Monimial, Bionomial, or Trinomial Name, or Taxonomic Concept Label) from Taxonomic Name Atoms
}
