% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{has_affiliation}
\alias{has_affiliation}
\title{Has Affiliation}
\format{An object of class \code{identifier} of length 4.}
\usage{
has_affiliation
}
\description{
Has Affiliation
}
\keyword{datasets}
