% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{has_content}
\alias{has_content}
\title{Creator (Author)}
\format{An object of class \code{identifier} of length 4.}
\usage{
has_content
}
\description{
Creator (Author)
}
\keyword{datasets}
