% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{has_keyword}
\alias{has_keyword}
\title{Has Keyword}
\format{An object of class \code{identifier} of length 4.}
\usage{
has_keyword
}
\description{
Has Keyword
}
\keyword{datasets}
