% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{has_publisher_id}
\alias{has_publisher_id}
\title{Has a Publisher}
\format{An object of class \code{identifier} of length 4.}
\usage{
has_publisher_id
}
\description{
Has a Publisher
}
\keyword{datasets}
