% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{has_scientific_name}
\alias{has_scientific_name}
\title{Has Scientific Name}
\format{An object of class \code{identifier} of length 4.}
\usage{
has_scientific_name
}
\description{
Has Scientific Name
}
\keyword{datasets}
