% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{has_secundum_string}
\alias{has_secundum_string}
\title{Secundum string}
\format{An object of class \code{identifier} of length 4.}
\usage{
has_secundum_string
}
\description{
Secundum string
}
\keyword{datasets}
