% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{has_taxonomic_rank}
\alias{has_taxonomic_rank}
\title{Dwc  Rank}
\format{An object of class \code{identifier} of length 4.}
\usage{
has_taxonomic_rank
}
\description{
Dwc  Rank
}
\keyword{datasets}
