% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{has_taxonomic_status}
\alias{has_taxonomic_status}
\title{Dwc Status}
\format{An object of class \code{identifier} of length 4.}
\usage{
has_taxonomic_status
}
\description{
Dwc Status
}
\keyword{datasets}
