% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{has_taxonomic_status_id}
\alias{has_taxonomic_status_id}
\title{Dwc IRI Status}
\format{An object of class \code{identifier} of length 4.}
\usage{
has_taxonomic_status_id
}
\description{
Dwc IRI Status
}
\keyword{datasets}
