% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{has_verbatim_rank}
\alias{has_verbatim_rank}
\title{Dwc Verbatim Rank}
\format{An object of class \code{identifier} of length 4.}
\usage{
has_verbatim_rank
}
\description{
Dwc Verbatim Rank
}
\keyword{datasets}
