% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{has_zoobank}
\alias{has_zoobank}
\title{Has Zoobank}
\format{An object of class \code{identifier} of length 4.}
\usage{
has_zoobank
}
\description{
Has Zoobank
}
\keyword{datasets}
