% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier_factory_new.R
\name{identifier_factory_new}
\alias{identifier_factory_new}
\title{Identifier factory}
\usage{
identifier_factory_new(fun, prefixes, def_prefix, schema_name, mongo_key)
}
\description{
A modified fn from rdf4r
}
