% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier_new.R
\name{identifier_new}
\alias{identifier_new}
\title{Identifier new}
\usage{
identifier_new(node, xml, schema_name, mongo_key, prefix = NA,
  blank = FALSE)
}
\description{
A modification of identifier() from rdf4r (queries mongo, instead of openbiodiv)
}
