% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxonomicNameUsage.R
\name{in_nomenclature_heading}
\alias{in_nomenclature_heading}
\title{Is the current XML node a child of nomenclature heading?}
\usage{
in_nomenclature_heading(node)
}
\arguments{
\item{node}{XML}
}
\description{
TODO This fucntion should probably get encapusalted inside a TNU
}
