% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambigation-person.R
\name{insert_sameas_person}
\alias{insert_sameas_person}
\title{Inserts all possible owl:sameAs properties for class Person}
\usage{
insert_sameas_person(person_id)
}
\description{
Inserts all possible owl:sameAs properties for class Person
}
