% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{institution_code_usage}
\alias{institution_code_usage}
\title{Institution Code Usage}
\usage{
institution_code_usage(atoms, identifiers, access_options)
}
\arguments{
\item{atoms}{a list of literals}

\item{identifiers}{a list of identifiers}

\item{access_options}{}
}
\value{
\code{ResourceDescriptionFramework}
}
\description{
This creates RDF for institutional_code usages in the document.
}
