% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-access-high-level.R
\name{institution_rule1}
\alias{institution_rule1}
\title{Rule1 for Institution Disambiguation}
\usage{
institution_rule1(affiliation)
}
\arguments{
\item{affiliation}{the affiliation string}
}
\value{
a URI with the match if found, NA if nothing is found
}
\description{
TODO unexport that for production
}
