% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-access-high-level.R
\name{institution_rule2}
\alias{institution_rule2}
\title{(2) look for an approxiamte match in the affiliation string:
(a) approximate match of the institution name and
(b) exact match of the city}
\usage{
institution_rule2(affiliation)
}
\arguments{
\item{affiliation}{the affiliation string}
}
\value{
URI or NA
}
\description{
(2) look for an approxiamte match in the affiliation string:
(a) approximate match of the institution name and
(b) exact match of the city
}
