% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{institutional_code}
\alias{institutional_code}
\title{Institutional Code}
\format{An object of class \code{identifier} of length 4.}
\usage{
institutional_code
}
\description{
Institutional Code
}
\keyword{datasets}
