% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{Introduction}
\alias{Introduction}
\title{Introduction:
  uri: http://purl.org/spar/deo/Introduction>
comment: an abstract of a science article}
\format{An object of class \code{identifier} of length 4.}
\usage{
Introduction
}
\description{
Introduction:
  uri: http://purl.org/spar/deo/Introduction>
comment: an abstract of a science article
}
\keyword{datasets}
