% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\name{is.AnonRDF}
\alias{is.AnonRDF}
\title{Is the object an Anonymous Triples List (RDF)?}
\usage{
is.AnonRDF(x)
}
\description{
Is the object an Anonymous Triples List (RDF)?
}
\seealso{
Other anonymous rdf: \code{\link{AnonRDF}}
}
\concept{anonymous rdf}
