% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nactem.R
\name{is.error}
\alias{is.error}
\title{Did you return an error?}
\usage{
is.error(e)
}
\arguments{
\item{e}{}
}
\value{
T/F
}
\description{
Did you return an error?
}
\examples{
is.error(simpleError("my bad"))

}
