% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\name{is.ResourceDescriptionFramework}
\alias{is.ResourceDescriptionFramework}
\title{Is the object an Triples List (RDF)?}
\usage{
is.ResourceDescriptionFramework(x)
}
\arguments{
\item{x}{object to check}
}
\value{
logical
}
\description{
Is the object an Triples List (RDF)?
}
\seealso{
Other rdf: \code{\link{ResourceDescriptionFramework}}
}
\concept{rdf}
