% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{is_contained_by}
\alias{is_contained_by}
\title{Creator (Author)}
\format{An object of class \code{identifier} of length 4.}
\usage{
is_contained_by
}
\description{
Creator (Author)
}
\keyword{datasets}
