% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{Journal}
\alias{Journal}
\title{a science journal}
\format{An object of class \code{identifier} of length 4.}
\usage{
Journal
}
\description{
a science journal
}
\keyword{datasets}
