% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdfization-keywords.R
\name{keyword_extractor}
\alias{keyword_extractor}
\title{Keyword Extractor}
\usage{
keyword_extractor(XML, metadata,
  XPath = yaml::yaml.load_file(obkms$config$keywords_db_xpath),
  keyword_scheme = obkms$parameters$Vocabularies$Subject_Classification_Terms)
}
\arguments{
\item{XML}{an XML object containing a keyword group}

\item{metadata}{a Taxonomic Article object containing the keyword group}

\item{XPath}{atom locations of a keyword group}

\item{keyword_scheme}{keyword group classifiaction scheme}
}
\value{
triples pertaining to the keyword
}
\description{
Keyword Extractor
}
