% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{keyword_group}
\alias{keyword_group}
\title{Keyword Group Constructor}
\usage{
keyword_group(atoms, identifiers, access_options, schema_name, mongo_key)
}
\arguments{
\item{atoms}{a list of literals}

\item{identifiers}{a list of identifiers}

\item{access_options}{}
}
\value{
\code{ResourceDescriptionFramework}
}
\description{
Keyword Group Constructor
}
