% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{KeywordGroup}
\alias{KeywordGroup}
\title{Keyword Group}
\format{An object of class \code{identifier} of length 4.}
\usage{
KeywordGroup
}
\description{
Keyword Group
}
\keyword{datasets}
