% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{last}
\alias{last}
\title{Last Element of Vector}
\usage{
last(x)
}
\value{
the last element of the vector
}
\description{
Last Element of Vector
}
\examples{
x = c(1, 2, 3)
y = c("a", "b", "c")
last(x)
last(y)

}
