% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{last_token}
\alias{last_token}
\title{Last Token of Character Vector}
\usage{
last_token(x, split, ...)
}
\arguments{
\item{x}{a character vector to extract last token from}

\item{split}{string or regex on which to do the split}

\item{...}{further arguments to strsplit}
}
\value{
last token of the character vector
}
\description{
Calls `strsplit` and then returns the last token.
}
\examples{
last_token("http://tb.plazi.org/GgServer/xml/uno", "/")
last_token(c("http://tb.plazi.org/GgServer/xml/dos", "http://tb.plazi.org/GgServer/xml/tres"), "/")

}
