% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxonomicName.R
\name{LatinName}
\alias{LatinName}
\title{Constructor for Latin name}
\usage{
LatinName(id, kingdom, phylum, class, order, family, genus, subgenus, species,
  subspecies, taxonomic_rank, taxonomic_status, verbatim_rank, verbatim_status,
  authorship)
}
\arguments{
\item{id}{URI, identifier of the name}

\item{kingdom}{character}

\item{phylum}{character}

\item{order}{character}

\item{family}{character}

\item{genus}{character}

\item{subgenus}{character}

\item{species}{character}

\item{subspecies}{character}

\item{taxonomic_status}{vocabulary item}

\item{verbatim_rank}{character}

\item{authorship}{character}

\item{rank}{vocabulary item}
}
\value{
LatinName
}
\description{
LatinName is a subclass of TaxonomicName.
}
