% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-access-high-level.R
\name{lookup_TaxonomicConcept_id}
\alias{lookup_TaxonomicConcept_id}
\title{Lookup the Identifier of the Taxonomic Concept Given the Identifier of its
  Taxonomic Concept Label}
\usage{
lookup_TaxonomicConcept_id(tcl_id)
}
\arguments{
\item{tcl_id}{Character. The Identifier of a taxonomic concept label.}
}
\value{
Character. The id of the taxonomic concept with the given label,
}
\description{
For the given TNU, locate (or create if non-existant) an URI in the system,
corresponding to the Latinized n name that TNU mentions.
}
