% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-access-high-level.R
\name{lookup_TaxonomicName_id}
\alias{lookup_TaxonomicName_id}
\title{Lookup Identifier of Taxonomic Name Used in a Taxonomic Name Usage}
\usage{
lookup_TaxonomicName_id(a_TaxonomicNameUsage, article_id,
  generate_on_fail = TRUE)
}
\arguments{
\item{article_id}{URI. The article (context) where the TNU is mentioned.}

\item{generate_on_fail}{Logical. What do to do if everything else fails.}
}
\value{
Character. The URI of
}
\description{
For the given TNU, locate (or create if non-existant) an URI in the system,
corresponding to the Latinized n name that TNU mentions.
}
\details{
TODO This function is overly complex, needs simplification
}
