% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-access-high-level.R
\name{lucene_author_query}
\alias{lucene_author_query}
\title{Lucene Query String to Search For Authors}
\usage{
lucene_author_query(first_name, last_name, email, affiliation)
}
\arguments{
\item{first_name}{the first name}

\item{last_name}{the last name}

\item{email}{author's email}
}
\value{
the lucene query string
}
\description{
Depending of what arguments are supplied, different queries are returned
(generic behavior)
}
