% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-access-high-level.R
\name{lucene_fuzzy_and_query}
\alias{lucene_fuzzy_and_query}
\title{Lucene Fuzzy AND Query}
\usage{
lucene_fuzzy_and_query(field_name, label, remove_stop = TRUE)
}
\arguments{
\item{field_name}{the field name to query against}

\item{label}{label to converted into a fuzzy and clause}
}
\value{
lucene query string
}
\description{
Lucene Fuzzy AND Query
}
\examples{

lucene_fuzzy_and_query( "last_name", "van Basten")

}
