% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-access-high-level.R
\name{lucene_label_query}
\alias{lucene_label_query}
\title{Lucene Label Query}
\usage{
lucene_label_query(label)
}
\arguments{
\item{label}{the label text that needs to be investigated}
}
\value{
the Lucene query string
}
\description{
Lucene Label Query
}
\examples{

lucene_label_query("Pensoft Publishers")

}
