% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-access-high-level.R
\name{lucene_phrase_query}
\alias{lucene_phrase_query}
\title{Query the Lucene Index for a Specific Phrase}
\usage{
lucene_phrase_query(phrase, field = "label",
  resource_type = "<http://www.w3.org/2002/07/owl#Thing>")
}
\arguments{
\item{phrase}{the exact phrase to look for (string)}

\item{field}{which field to query, default is label}

\item{resource_type}{the URI of the class to which you want to restrict your queries}
}
\value{
a list of hits and scores
}
\description{
TODO need lucene prefix
}
