% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-access-high-level.R
\name{lucene_query_creator.fuzzy_field}
\alias{lucene_query_creator.fuzzy_field}
\title{Lucene Query Creator: Fuzzy Label Query}
\usage{
lucene_query_creator.fuzzy_field(field_name, label)
}
\arguments{
\item{label}{the label to be parsed}
}
\value{
the query string and the needed connector (a vector of two)
}
\description{
Takes a string, a creates a fuzzy label query out of it.
}
