% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{mentionsHabitat}
\alias{mentionsHabitat}
\title{mentions habitat}
\format{An object of class \code{identifier} of length 4.}
\usage{
mentionsHabitat
}
\description{
mentions habitat
}
\keyword{datasets}
