% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{metadata}
\alias{metadata}
\title{Metadata constructor}
\usage{
metadata(atoms, identifiers, prefix, schema_name, mongo_key)
}
\arguments{
\item{atoms}{a list of literals}

\item{identifiers}{a list of identifiers}

\item{prefix}{}
}
\value{
\code{ResourceDescriptionFramework}
}
\description{
Metadata constructor
}
