% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{new_pensoft_articles}
\alias{new_pensoft_articles}
\title{New Pensoft Articles}
\usage{

  new_pensoft_articles(endpoint = "http://bdj.pensoft.net/lib/journal_archive.php?journal=bdj",
  fromdate = as.Date("2010-01-01"))
}
\arguments{
\item{fromdate}{from which date on. If missing starts from 1.jan.2010. Date needs to be a Date object}

\item{journal}{endpoint from which to dump. if missing dumps BDJ}
}
\value{
ZIP file as a binary object that can be written in the filesystem
}
\description{
New Pensoft Articles
}
