% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{Nomenclature}
\alias{Nomenclature}
\title{Nomenclature
 the informational content of an article}
\format{An object of class \code{identifier} of length 4.}
\usage{
Nomenclature
}
\description{
Nomenclature
 the informational content of an article
}
\keyword{datasets}
