% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{NomenclatureCitation}
\alias{NomenclatureCitation}
\title{Nom Citation
the informational content of an article}
\format{An object of class \code{identifier} of length 4.}
\usage{
NomenclatureCitation
}
\description{
Nom Citation
the informational content of an article
}
\keyword{datasets}
