% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{NomenclatureCitationsList}
\alias{NomenclatureCitationsList}
\title{Nom Citations
the informational content of an article}
\format{An object of class \code{identifier} of length 4.}
\usage{
NomenclatureCitationsList
}
\description{
Nom Citations
the informational content of an article
}
\keyword{datasets}
