% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{Organization}
\alias{Organization}
\title{Organization}
\format{An object of class \code{identifier} of length 4.}
\usage{
Organization
}
\description{
Organization
}
\keyword{datasets}
