% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-zenodo.R
\name{params2string}
\alias{params2string}
\title{Converts a List of Parameters to a URL GET String}
\usage{
params2string(...)
}
\arguments{
\item{...}{named list of character(1)}
}
\value{
character(1) the parameters converted to a string useful for GET
}
\description{
Converts a List of Parameters to a URL GET String
}
\examples{
params2string(page = 1, size = 10)

}
