% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{parent_id}
\alias{parent_id}
\title{Get the Parent OBKS Id for an XML Node}
\usage{
parent_id(node, fullname = FALSE)
}
\arguments{
\item{node}{the XML node from which to take the ID; cannot be missing!}

\item{fullname}{if TRUE, returns a URI with the OBKMS base prefix}
}
\description{
Does not do any database lookups. Recursive looks for the parent node until
it finds one with ID.
}
