% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{pasteif}
\alias{pasteif}
\title{Paste If}
\usage{
pasteif(..., cond)
}
\arguments{
\item{...}{one or more R objects, to be passed to paste0 if cond is TRUE.}

\item{cond}{boolean, condition to be true.}
}
\value{
string
}
\description{
Pastes something only if the condition is met. Uses paste0.
}
