% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pensoft.R
\name{pensoft_article_eloc}
\alias{pensoft_article_eloc}
\title{Pensoft Article E-Location}
\usage{
pensoft_article_eloc(articles)
}
\arguments{
\item{articles}{character(n) of links to article XML's}
}
\value{
character(n) eloction
}
\description{
Should work with any Taxpub XML!
}
\examples{
pensoft_article_eloc("/media/obkms/pensoft-corpus.xml/10.3897_BDJ.2.e4168.xml")

}
