% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pensoft.R
\name{pensoft_images}
\alias{pensoft_images}
\title{Pensoft Links for Images of a Given Pensoft Paper}
\usage{
pensoft_images(articles)
}
\arguments{
\item{articles}{character(n) of links to article XML's}
}
\value{
character(n) of links to images contained in the article
}
\description{
Should work with any Taxpub XML!
}
\examples{
pensoft_images("/media/obkms/pensoft-corpus.xml/10.3897_BDJ.2.e4168.xml")

}
