% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambigation-person.R
\name{person_constructor}
\alias{person_constructor}
\title{Person Constructor}
\usage{
person_constructor(id)
}
\arguments{
\item{id}{}
}
\value{
an object of class Person
}
\description{
Person Constructor
}
