% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{PersonalIdentifier}
\alias{PersonalIdentifier}
\title{Personal Identifier}
\format{An object of class \code{identifier} of length 4.}
\usage{
PersonalIdentifier
}
\description{
Personal Identifier
}
\keyword{datasets}
