% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-access-high-level.R
\name{phrase2FuzzyAnd}
\alias{phrase2FuzzyAnd}
\title{Convert a Literal Phrase to a Fuzzy-AND-Phrase}
\usage{
phrase2FuzzyAnd(phrase)
}
\value{
improved phrase
}
\description{
Puts qunituple quotes around the phrase, removes special
characters, adds tilde after the words and AND between
the words of a phrase
}
