% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plazi.R
\name{plazi_feed}
\alias{plazi_feed}
\title{Plazi Feed Constructor}
\usage{
plazi_feed(feed_url = "http://tb.plazi.org/GgServer/xml.rss.xml")
}
\arguments{
\item{feed_url}{string, the URL of the Plazi feed. A default is present.}
}
\value{
plazi_feed, xml_document, the Plazi feed, as a local object.
}
\description{
Constructs a Plazi Feed Object from the Plazi Feed URL
}
\examples{
todays_plazi_feed = plazi_feed()
xml2::write_xml(todays_plazi_feed, file = "~/tmp/plazi_feed.xml")
todays_plazi_feed = plazi_feed("~/tmp/plazi_feed.xml")

}
