% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plazi.R
\name{plazi_images}
\alias{plazi_images}
\title{Retrieve Links for Images of a Given Treatment}
\usage{
plazi_images(id, directory)
}
\arguments{
\item{directory}{}
}
\value{
character(n) of links to images
}
\description{
Retrieve Links for Images of a Given Treatment
}
\examples{
plazi_images("http://tb.plazi.org/GgServer/xml/F58CB27430C0CCBD272770124C31BFF8")

}
