% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonx.R, R/taxpub.R
\docType{data}
\name{plazi_int}
\alias{plazi_int}
\alias{plazi_int}
\title{Plazi Internal}
\format{An object of class \code{xml_schema} (inherits from \code{R6}) of length 13.}
\usage{
plazi_int

plazi_int
}
\description{
Plazi Internal

Plazi Internal
}
\keyword{datasets}
