% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{pref_label}
\alias{pref_label}
\title{SKOS Preferred Label}
\format{An object of class \code{identifier} of length 4.}
\usage{
pref_label
}
\description{
SKOS Preferred Label
}
\keyword{datasets}
