% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentComponent.R
\name{print.DocumentComponent}
\alias{print.DocumentComponent}
\title{Print Method for DocumentComponent}
\usage{
\method{print}{DocumentComponent}(obj)
}
\arguments{
\item{obj}{DocumentComponent to print}
}
\description{
Print Method for DocumentComponent
}
