% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxonomicName.R
\name{print.TaxonomicName}
\alias{print.TaxonomicName}
\title{Print Method for TaxonomicName}
\usage{
\method{print}{TaxonomicName}(obj)
}
\arguments{
\item{obj}{TaxonomicName}
}
\description{
Print Method for TaxonomicName
}
