% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{processing_status}
\alias{processing_status}
\title{Find (Re-)Processing Status}
\usage{
processing_status(node)
}
\arguments{
\item{node}{\code{XML2} object}
}
\value{
\code{logical} TRUE if the node has been processed, FALSE otherwise
}
\description{
Find (Re-)Processing Status
}
