% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{publication_date}
\alias{publication_date}
\title{publication_date}
\format{An object of class \code{identifier} of length 4.}
\usage{
publication_date
}
\description{
publication_date
}
\keyword{datasets}
