% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparql.R
\docType{data}
\name{qlookup_author1}
\alias{qlookup_author1}
\title{Looks up an author's ID by his or her email.}
\format{An object of class \code{character} of length 1.}
\usage{
qlookup_author1
}
\description{
Idea: first letter of first name AND
      exact match of email AND
      fuzzy match of last name
}
\keyword{datasets}
