% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparql.R
\docType{data}
\name{qlookup_by_label_in_context}
\alias{qlookup_by_label_in_context}
\title{Look Up Anything by Label In Context}
\format{An object of class \code{character} of length 1.}
\usage{
qlookup_by_label_in_context
}
\description{
Most general lookup by label query.
}
\section{Fields}{

\describe{
\item{\code{label}}{}
}}

\keyword{datasets}
