% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparql.R
\docType{data}
\name{qlookup_by_label_no_context}
\alias{qlookup_by_label_no_context}
\title{Look Up Anything by Label No Context}
\format{An object of class \code{character} of length 1.}
\usage{
qlookup_by_label_no_context
}
\description{
Equivalent to lookup_by_label but with an additional context argument
ignored.
opebiodiv_id_via_label_lookup
Most general lookup by label query.
}
\section{Fields}{

\describe{
\item{\code{label}}{}
}}

\keyword{datasets}
