% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparql.R
\docType{data}
\name{qlookup_taxonomic_concept_id}
\alias{qlookup_taxonomic_concept_id}
\title{Lookup the Taxonomic Concept of a Treatment}
\format{An object of class \code{character} of length 1.}
\usage{
qlookup_taxonomic_concept_id
}
\description{
Looks for an object of type paper insde the article context.
}
\section{Fields}{

\describe{
\item{\code{article_id}}{}
}}

\keyword{datasets}
