% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\docType{data}
\name{RdfLibBackend}
\alias{RdfLibBackend}
\title{RDF Class with \code{rdflib} backend}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
RdfLibBackend
}
\description{
RDF Class with \code{rdflib} backend
}
\keyword{datasets}
