% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{realization}
\alias{realization}
\title{Realization Of}
\format{An object of class \code{identifier} of length 4.}
\usage{
realization
}
\description{
Realization Of
}
\keyword{datasets}
