% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{regularize_taxonomic_names}
\alias{regularize_taxonomic_names}
\title{Regularize Taxonomic Names in an XML Document}
\usage{
regularize_taxonomic_names(xmldoc, schema)
}
\arguments{
\item{xmldoc}{the taxonomic article as a XML2 document}

\item{schema}{a list with mark-up information for taxonomic names in the XML schema}
}
\value{
a vector of regularized name parts
}
\description{
Side-effect: modifies the XML object by regularizing taxonimic names!
}
