% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nactem.R
\name{regularize_taxonomic_names.taxonx}
\alias{regularize_taxonomic_names.taxonx}
\title{Regularize Plazi XML}
\usage{
\method{regularize_taxonomic_names}{taxonx}(xmldoc, TNU_markup, PTNU_markup,
  regPTNU_markup)
}
\arguments{
\item{xmldoc}{}
}
\value{
a vector of regularized name parts
}
\description{
TODO: generic function
}
