% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nactem.R
\name{regularize_taxonomic_names.taxpub}
\alias{regularize_taxonomic_names.taxpub}
\title{Regularize Taxonomic Names in an XML Document (TaxPub)}
\usage{
\method{regularize_taxonomic_names}{taxpub}(xmldoc, TNU_markup, PTNU_markup,
  regPTNU_markup)
}
\arguments{
\item{xmldoc}{the taxonomic article as a XML2 document}

\item{PTNU_markup}{how taxonomic name parts/ partial taxonomic name usages are marked up}

\item{regPTNU_markup}{how regularized parts are marked up}
}
\value{
a vector of regularized name parts
}
\description{
Side-effect: modifies the XML object by regularizing taxonimic names!
}
