% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{remove_all_tags}
\alias{remove_all_tags}
\title{Strips all XML tags}
\usage{
remove_all_tags(xmldoc)
}
\arguments{
\item{xmldoc}{}
}
\value{
a vector of the document as plain text
}
\description{
Strips all XML tags
}
