% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nactem.R
\name{remove_plazi_metadata}
\alias{remove_plazi_metadata}
\title{Remove Plazi Metadata from Taxonomic Names}
\usage{
remove_plazi_metadata(node, metadata_markup = "tax:xmldata")
}
\arguments{
\item{TNU_markup}{}
}
\value{
the new text of the taxonomic name usages
}
\description{
Side-effect: modifies the XML doc
}
