% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nactem.R
\name{remove_taxpub_markup}
\alias{remove_taxpub_markup}
\title{Remove TaxPub markup (Generic)}
\usage{
remove_taxpub_markup(xmldoc, ...)
}
\arguments{
\item{...}{}
}
\description{
Modifies the XML object by removing superflous Taxpub markup
}
