% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\docType{data}
\name{ResourceDescriptionFramework}
\alias{ResourceDescriptionFramework}
\title{Mutable RDF Object}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ResourceDescriptionFramework
}
\arguments{
\item{initialize(size=100)}{Constructor. Use this to create new objects
it has a default tuning parameter. You may want to set it to the average
number of triples per processed document.}

\item{set_context(context)}{Context needs to be an \code{identifier}
object that correspond to the named graph of where the statements
are stored.}

\item{get_prefixes()}{Returns the prefixes as a deduplicated named
character vector of prefix to namespace mappings.}

\item{add_triple(subject, predicate, object)}{\itemize{
  \item{\code{subject} \code{identifier}}
  \item{\code{predicate} \code{identifier}}
  \item{\code{object} \code{identifier} or \code{literal} or \code{AnonymousRDF}.
  Use the anonymous RDF if you hop through a blank node.}
} Returns the success status (\code{logical})}

\item{add_triples(ll)}{ll needs to be a \code{ResourceDescriptionFramework}
object. The information is merged.}

\item{serialize()}{Returns the Turtle serialization.}

\item{prefix_list}{\code{dynamic_vector} Prefixes of all the stored
identifiers as an uncollapsed list. For most cases you would want
to use \code{get_prefixes}.}

\item{context}{\code{identifier}. The named graph of where the statements
are stored.}
}
\description{
This is a mutable object. Adds, stores, and serializes RDF statements
efficiently. Uses an amortized vector \code{DynVector} as a internal
storage. You can inherit this class and override to create your own
implementation and still retain syntax compatibility.
}
\seealso{
Other rdf: \code{\link{is.ResourceDescriptionFramework}}
}
\concept{rdf}
\keyword{datasets}
