% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{ResourceIdentifier}
\alias{ResourceIdentifier}
\title{Resource Identifier}
\format{An object of class \code{identifier} of length 4.}
\usage{
ResourceIdentifier
}
\description{
Resource Identifier
}
\keyword{datasets}
