% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdfization-metadata.R
\name{retrieve_affiliation_string_from_author}
\alias{retrieve_affiliation_string_from_author}
\title{Help function
TODO shouldn't be exported}
\usage{
retrieve_affiliation_string_from_author(reference_id, xml_document,
  affiliation_xpath = "/article/front/article-meta/aff[@id='\%refid']")
}
\description{
Help function
TODO shouldn't be exported
}
