% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{Role}
\alias{Role}
\title{a role of agent in time}
\format{An object of class \code{identifier} of length 4.}
\usage{
Role
}
\description{
a role of agent in time
}
\keyword{datasets}
