% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropenbio.R
\docType{package}
\name{ropenbio}
\alias{ropenbio}
\alias{ropenbio-package}
\title{ropenbio: R Package Accompanying the Open Biodiversity Knowledge Management
System}
\description{
The functions in this package are separated into the following categories:
}
\section{Low-level database access functions}{

Provide access to the database at level of triples, repositories, queries,
etc.
}

\section{High-level database access functions}{

Provide access to the database at the level of entities of OpenBiodiv such as
papers, authors, taxa.
}

\section{Disambiguation procedure for persons}{

Run this procedure after RDF- #' ization to create \code{owk:sameAs}
relations between potentially congruent persons.
}

\section{Initialization procedure}{

Needs to be run before using the package to set-up certain global variables
shared accross all the functions.
}

\section{Logging procedures}{

}

\section{RDF-ization functions for author records}{

}

\section{RDF-ization functions for metadata}{

}

\section{Functions to serialize RDF}{

}

