% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-access-high-level.R
\name{rule_based_lookup}
\alias{rule_based_lookup}
\title{Generic Lookup Function}
\usage{
rule_based_lookup(parameter.lst, rule)
}
\arguments{
\item{rule}{a character vector of rules to be applied in sequential order until
one succeeds}

\item{lookup_parameter.lst}{a list of lookup parameters to be passed to the rules}
}
\description{
Generic Lookup Function
}
