% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambigation-person.R
\name{same_as_query}
\alias{same_as_query}
\title{owl:sameAs SPARQL query}
\usage{
same_as_query(id1, id2)
}
\arguments{
\item{id1}{the first ID, must a vector long exactly 1}

\item{id2}{the second ID, must be a vector long at least 1}
}
\value{
SPARQL query string where id1 is the same as all of the id2's,
  unprefixed
}
\description{
owl:sameAs SPARQL query
}
