% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambigation-person.R
\name{same_person}
\alias{same_person}
\title{Same Person}
\usage{
same_person(id)
}
\arguments{
\item{id}{an person URI in OpenBiodiv}
}
\value{
all other URI's that point to the same person
}
\description{
- construct the person object for the person URI
- run all three rules
}
