% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-access-low-level.R
\name{server_access_options.file}
\alias{server_access_options.file}
\title{Server Access Options Constructor from File}
\usage{
server_access_options.file(filename)
}
\arguments{
\item{filename}{a YAML file containing the server access options}
}
\value{
a \code{Server_Access_Options} object read from the file
}
\description{
Additionally, it reads the access secret (password) from the environment
variable \code{OBKMS_SECRET}
}
\details{
TODO: rename to server_access_options.file
TODO: class names ought to be a global parameter (like Server_Access_Options)
}
