% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{set_obkms}
\alias{set_obkms}
\title{NOT USED ANYMORE
Set the OBKMS ID of an XML node according to new id format}
\usage{
set_obkms(schema_name, mongo_name)
}
\arguments{
\item{schema_name}{the name of the xml_schema}

\item{mongo_name}{the names of mongo_key}
}
\value{
the local id (not an identifier object)
}
\description{
Does not do any database lookups.
}
