% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialization-rdf.R
\name{squote}
\alias{squote}
\title{Quote Characters to USE in RDF Serializations}
\usage{
squote(literal, language, literal_type)
}
\arguments{
\item{literal}{Character. The string that needs to be quoted.}

\item{language}{Character with the language code or list with the code as
attribute named `semantic_code`}

\item{literal_type}{Character. The type. Look at the webpage in the description.
If string, just leave missing.}
}
\description{
Information:
  http://iswc2011.semanticweb.org/fileadmin/iswc/Papers/Workshops/SSWS/Emmons-et-all-SSWS2011.pdf
}
\examples{

English = obkms$parameters$Language$English
squote("Pensoft Publishers", language = English )

Year = obkms$paramters$literal_type$Year
squote("2017", Year)

}
