% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{standard_injector}
\alias{standard_injector}
\title{Standard injector. Sets the id from the XPATH}
\usage{
standard_injector(obkms_id, xml_node)
}
\arguments{
\item{xml_node}{}
}
\description{
Standard injector. Sets the id from the XPATH
}
