% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{strip_filename_extension}
\alias{strip_filename_extension}
\title{Strip Extension From a Filename}
\usage{
strip_filename_extension(filename)
}
\arguments{
\item{filename}{filename}
}
\value{
the filename without the extension
}
\description{
Strip Extension From a Filename
}
