% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{strip_trailing_slash}
\alias{strip_trailing_slash}
\title{Strip Trailing Slash If Present}
\usage{
strip_trailing_slash(x)
}
\arguments{
\item{x}{character vector.}
}
\value{
x without the trailing slash, if present
}
\description{
Works only on UNIX.
}
\examples{
strip_trailing_slash("/media/obkms/plazi-corpus-xml/")


}
