% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{taxonomic_name_usage}
\alias{taxonomic_name_usage}
\title{Taxonomic Name Usage Constructor}
\usage{
taxonomic_name_usage(atoms, identifiers, access_options)
}
\arguments{
\item{atoms}{a list of literals}

\item{identifiers}{a list of identifiers}

\item{access_options}{}
}
\value{
\code{ResourceDescriptionFramework}
}
\description{
Taxonomic Name Usage Constructor
}
