% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic-rank.R
\name{taxonomic_rank}
\alias{taxonomic_rank}
\title{Maps Verbatim Rank String to a Vocabulary of Taxonomic Ranks}
\usage{
taxonomic_rank(verbatim_rank)
}
\arguments{
\item{verbatim_rank}{(character) the rank as a free-form string}
}
\value{
a vocabulary item (`TaxonomicRank`) corresponding to the rank encoded
}
\description{
Constructor for taxonomic rank
}
