% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic-status.R
\name{taxonomic_status}
\alias{taxonomic_status}
\title{Maps Verbatim Status String to a Vocabulary of Taxonomic Ranks}
\usage{
taxonomic_status(verbatim_status)
}
\arguments{
\item{verbatim_status}{(character) the status as a free-form string}
}
\value{
a vocabulary item (taxonomic status) corresponding to the rank encoded
}
\description{
Constructor for taxonomic status
}
