% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{TaxonomicArticle_extractor}
\alias{TaxonomicArticle_extractor}
\title{Top Level (Metadata) Extractor for a Taxnomic Article}
\usage{
TaxonomicArticle_extractor(xml,
  xlit = yaml::yaml.load_file(obkms$config$literals_db_xpath),
  xdoco = yaml::yaml.load_file(obkms$config$non_literals_db_xpath))
}
\arguments{
\item{xml}{an XML document object (as parsed by the `xml2` package) containg
a taxonomic paper}

\item{xlit}{a vector of XPATH locations of the atoms (literals) of the metadata
of a taxonomic document}

\item{xdoco}{a vector of XPATH locations of where other objects are to be
found which need to be processed recursively and attached to the paper}
}
\value{
triples
}
\description{
First this function finds all of the atoms that are belong to the metadata
of a taxonomic paper and creates triples for it. Then, it chops up the paper
XML into smaller parts that are sent to lower-level extractors to get
triples for them. After all of this everything is concattenated and returned.
}
\details{
TODO Note: this function does a little more than it should: it is also a
constructor
and rdfi-zer of TaxonomicArticle. Also it works with the paper object.
}
