% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxonomicName.R
\name{TaxonomicConceptLabel}
\alias{TaxonomicConceptLabel}
\title{Constructor for TaxonomicConceptLabel}
\usage{
TaxonomicConceptLabel(id, kingdom, phylum, class, order, family, genus,
  subgenus, species, subspecies, taxonomic_rank, taxonomic_status,
  verbatim_rank, verbatim_status, authorship, secundum_literal, secundum)
}
\arguments{
\item{id}{URI, identifier of the name}

\item{kingdom}{character}

\item{phylum}{character}

\item{order}{character}

\item{family}{character}

\item{genus}{character}

\item{subgenus}{character}

\item{species}{character}

\item{subspecies}{character}

\item{verbatim_rank}{character}

\item{authorship}{character}

\item{secundum_literal}{character}

\item{secundum}{URI}

\item{rank}{vocabulary item}

\item{status}{vocabulary item}
}
\value{
TaxonomicConceptLabel
}
\description{
TaxonomicConceptLabel is a subclass of LatinName containing s sec./sensu
part.
}
