% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{TaxonomicNameUsage}
\alias{TaxonomicNameUsage}
\title{TNU}
\format{An object of class \code{identifier} of length 4.}
\usage{
TaxonomicNameUsage
}
\description{
TNU
}
\keyword{datasets}
