% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{TaxonomicNameUsage_extractor}
\alias{TaxonomicNameUsage_extractor}
\title{Taxonomic Name Usage Extractor}
\usage{
TaxonomicNameUsage_extractor(comp, metadata)
}
\arguments{
\item{comp}{an XML componennt containing a taxonomic name usage}

\item{metadata}{The metadata of the article. Needed to get the context.}
}
\value{
triples of RDF
}
\description{
Extracts knowledge from an XML node containing a taxonomic name usage.
Extracts also names.
}
