% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{Title}
\alias{Title}
\title{Title:
uri: <http://purl.org/spar/doco/Title>
  comment: the title of an article}
\format{An object of class \code{identifier} of length 4.}
\usage{
Title
}
\description{
Title:
uri: <http://purl.org/spar/doco/Title>
  comment: the title of an article
}
\keyword{datasets}
