% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Treatment.R
\name{Treatment}
\alias{Treatment}
\title{Constructor for Treatment}
\usage{
Treatment(node)
}
\arguments{
\item{node}{XML2 node containing a node.}
}
\value{
a Treatment object
}
\description{
Constructor for Treatment
}
