% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{Treatment_extractor}
\alias{Treatment_extractor}
\title{Extracts RDF from a Treatment Node}
\usage{
Treatment_extractor(node, xml_schema)
}
\arguments{
\item{node}{XML node. Contains a Treatment.}

\item{xml_schema}{character, contains the XML name of the XML schema to use}
}
\value{
RDF
}
\description{
Extracts RDF from a Treatment Node
}
