% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plazi.R
\name{treatment_language}
\alias{treatment_language}
\title{Get Plazi Treatment Language}
\usage{
treatment_language(id, directory)
}
\arguments{
\item{id}{a character vector of Plazi treatment ID's.}

\item{directory}{a string, location of treatments.}
}
\value{
a named character vector where each entry corresponds to an
  individual treatment's language.
}
\description{
This is a parallelized funciton.
}
\examples{

lang = treatment_language(sort(treatments$id)[1:40], directory = "/home/viktor2/tmp")

}
