% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialization-rdf.R
\name{triple2}
\alias{triple2}
\title{Triple Constructor}
\usage{
triple2(S, P, O, blank = FALSE)
}
\arguments{
\item{S}{Subject}

\item{P}{Predicate}

\item{O}{Object}

\item{blank}{Logical: if it is a blank (anononymous node)}
}
\value{
NULL if something went wrong, a triple object otherwise.
  (if the object is a resource or a literal), or a list (if the object is
  a list).
}
\description{
A triple is a single sentence in RDF. The subject of a triple can be a
blank node ("") or a resource (an IRI encoded as character). The predicate
of the triple must necessarily be a resource. The object of the triple can
be a resource, a literal (character), or a list of triples,
all of which are blank.
}
