% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{underscore_taxonomic_names}
\alias{underscore_taxonomic_names}
\title{Underscore Taxonimic Names in an XML Document Object}
\usage{
underscore_taxonomic_names(xmldoc, schema)
}
\arguments{
\item{xmldoc}{an XML2 object in which to underscore the names}

\item{schema}{the properties of the XML schema as a list}
}
\value{
a vector of underscored names
}
\description{
Side-effect: modifies the XML object by underscoring taxonimic names
}
