% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{unescape_html}
\alias{unescape_html}
\title{Unescpae HTML Characters}
\usage{
unescape_html(str)
}
\arguments{
\item{str}{input string}
}
\value{
string without the escape characters
}
\description{
Unescpae HTML Characters
}
