% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_repository}
\alias{update_repository}
\title{Perform update described in a SPARQL 1.1 Update string
Based on http://docs.rdf4j.org/rest-api/#_perform_update_described_in_a_sparql_1_1_update_string}
\usage{
update_repository(server_access_options, repository, update)
}
\arguments{
\item{server_access_options}{\emph{list} of rdf4j server access options}

\item{repository}{\emph{character} id of the repository to be updated}

\item{update}{\emph{character} the update string}
}
\value{
success status of the operation
}
\description{
Perform update described in a SPARQL 1.1 Update string
Based on http://docs.rdf4j.org/rest-api/#_perform_update_described_in_a_sparql_1_1_update_string
}

