% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biology.R
\name{verbstat2openbiodiv}
\alias{verbstat2openbiodiv}
\title{Converts a Verbatim Status to a Controlled Vocabulary}
\usage{
verbstat2openbiodiv(verbatim_status,
  vocab_file = paste0(path.package("ropenbio"), "/inst/vocabulary/",
  "taxonomic-status.yml"), def_prefix)
}
\arguments{
\item{verbatim_status}{string}
}
\value{
identifier from OpenBiodiv
}
\description{
Converts a Verbatim Status to a Controlled Vocabulary
}
