% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{Work}
\alias{Work}
\title{a creative or scientific work}
\format{An object of class \code{identifier} of length 4.}
\usage{
Work
}
\description{
a creative or scientific work
}
\keyword{datasets}
