% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{xml2nactem}
\alias{xml2nactem}
\title{Convert an XML File to Plain Text NaCTeM Format}
\usage{
xml2nactem(xml_filename, xml_schema, modify_xml = FALSE)
}
\arguments{
\item{xml_filename}{the filename (one file) to convert to NaCTeM plain text}

\item{xml_schema}{a list with mark-up information for taxonomic names in the XML schema}

\item{modify_xml}{if TRUE, the function has a side-effect and modifies the XML on disk}
}
\value{
plain text to write as NaCTeM format
}
\description{
Possible side effect: file gets modified on disk!
}
