% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_schema.R
\docType{data}
\name{XmlSchema}
\alias{XmlSchema}
\title{R6 Class for XML Schemas}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
XmlSchema
}
\description{
R6 Class for XML Schemas
}
\section{Fields}{

\describe{
\item{\code{schema_name}}{the name of the schema}

\item{\code{file_pattern}}{regular expression pattern matching the file-name extension}

\item{\code{extension}}{the file-name extension}

\item{\code{prefix}}{the prefix that these documents have in the URI}
}}

\examples{
taxonx = XmlSchema$new(schema_name = "taxonx", file_pattern = ".*\\\\.xml")

}
\keyword{datasets}
