% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr-zenodo.R
\name{zenodeo_fetch_ids}
\alias{zenodeo_fetch_ids}
\title{Fetches BLR Record Id's From Zenodo via Zenodeo}
\usage{
zenodeo_fetch_ids(start = 1, offset = 10, all = FALSE,
  endpoint = "http://zenodeo.punkish.org/v1/", ...)
}
\arguments{
\item{start}{integer(1) where to start from}

\item{offset}{integer(1) how many records to fetch}

\item{all}{logical(1) if true will keep iterating through the API until the end is reached/ fetches all id's; default behaviour is to stop after fetching one page}

\item{endpoint}{character(1) endpoint address with "/" at the end}

\item{...}{further paramters to be passed to the Zenodeo "records" API}
}
\value{
vector('list', n) of BLR id's.
}
\description{
Access the Zenodo API via a Zenodeo instance stored at endpoint.
}
\examples{
zenodeo_fetch_ids(1, 10)

}
