% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{author}
\alias{author}
\title{Author constructor}
\usage{
author(atoms, identifiers, prefix, new_taxons, mongo_key)
}
\arguments{
\item{atoms}{a list of literals}

\item{identifiers}{a list of identifiers}

\item{prefix}{}
}
\value{
\code{ResourceDescriptionFramework}
}
\description{
TODO rewrite the atoms as individual arguments. I would need a
do.call in the extractor for that.
}
