% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{checklist}
\alias{checklist}
\title{Checklist Section Constructor}
\usage{
checklist(atoms, identifiers, prefix, new_taxons, mongo_key)
}
\arguments{
\item{atoms}{a list of literals}

\item{identifiers}{a list of identifiers}

\item{prefix}{}
}
\value{
\code{ResourceDescriptionFramework}
}
\description{
Checklist Section Constructor
}
