% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{distribution}
\alias{distribution}
\title{Distribution Section Constructor}
\usage{
distribution(atoms, identifiers, prefix, new_taxons, mongo_key)
}
\arguments{
\item{atoms}{a list of literals}

\item{identifiers}{a list of identifiers}

\item{prefix}{}
}
\value{
\code{ResourceDescriptionFramework}
}
\description{
Distribution Section Constructor
}
