% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{find_literals}
\alias{find_literals}
\title{Finds the Atoms in a XML Node}
\usage{
find_literals(xml, xml_schema)
}
\arguments{
\item{xml}{the XML node}

\item{xpath}{the atom locations as a named character vector}

\item{atom_types}{the type (explicitly stated, not as xpath) of the atom}

\item{atom_lang}{the language (as xpath), if the xpath fails, will set to
en (if string)}
}
\value{
list
}
\description{
Finds the Atoms in a XML Node
}
