% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repository_information.R
\name{get_protocol_version}
\alias{get_protocol_version}
\title{Get the communication protocol version}
\usage{
get_protocol_version(options)
}
\arguments{
\item{options}{a list returned by \code{create_server_options}.}
}
\value{
a string, the protocol version.
}
\description{
This function queries the RDF endpoint for information about the protocol version.
}
\examples{
\dontrun{get_protocol_version(options)}

}

