% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{introduction_section}
\alias{introduction_section}
\title{Introduction Section Constructor}
\usage{
introduction_section(atoms, identifiers, prefix, new_taxons, mongo_key,
  publisher_id, journal_id, plazi_doc)
}
\arguments{
\item{atoms}{a list of literals}

\item{identifiers}{a list of identifiers}

\item{prefix}{}
}
\value{
\code{ResourceDescriptionFramework}
}
\description{
Introduction Section Constructor
}
