% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{node_extractor}
\alias{node_extractor}
\title{Node Extractor}
\usage{
node_extractor(node, xml_schema, reprocess, triples, access_options,
  dry = FALSE, filename, root_id = root_ident)
}
\arguments{
\item{xml_schema}{XML schema of the XML resource; one of
taxpub taxonx or plazi_int}

\item{reprocess}{\code{logical}}

\item{triples}{the RDF object where the newly created triples will be
stored (note this is a side-effect/pass by reference)}

\item{access_options}{endpoint to database to which to submit}

\item{filename}{locator, e.g. file path or URL, of the XML resource}
}
\value{
\code{ResourceDescriptionFramework} containing all the triples from
  the article
}
\description{
Given an XML2 object, and an \code{ResourceDescriptionFramework} object, which
are both mutable, the \code{ResourceDescriptionFramework} object is
mutated into containing the triples extracted from the XML2 object,
while the XML2 object is possibly mutated into containing some annotations
allowing the faster reprocessing.
}
