% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonx.R
\docType{data}
\name{plazi_feed_schema}
\alias{plazi_feed_schema}
\title{Plazi Feed}
\format{An object of class \code{xml_schema} (inherits from \code{R6}) of length 13.}
\usage{
plazi_feed_schema
}
\description{
Plazi Feed
}
\keyword{datasets}
