% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_schema.R
\docType{data}
\name{plazi_int}
\alias{plazi_int}
\title{Plazi Internal}
\format{An object of class \code{xml_schema} (inherits from \code{R6}) of length 9.}
\usage{
plazi_int
}
\description{
Plazi Internal
}
\keyword{datasets}
