% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nactem.R
\name{regularize_taxonomic_names}
\alias{regularize_taxonomic_names}
\title{Regularize Taxonomic Names in an XML Document (Generic)}
\usage{
regularize_taxonomic_names(xmldoc, ...)
}
\arguments{
\item{...}{}
}
\value{
a vector of regularized name parts
}
\description{
Regularize Taxonomic Names in an XML Document (Generic)
}
