% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{root_id}
\alias{root_id}
\title{Get the Root (Article) OBKMS Id for an XML Node}
\usage{
root_id(node, fullname = FALSE)
}
\arguments{
\item{fullname}{}
}
\description{
Does not do any database lookups - only looks at XML
}
