% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{taxonomic_key}
\alias{taxonomic_key}
\title{Taxonomic Key Section Constructor}
\usage{
taxonomic_key(atoms, identifiers, access_options)
}
\arguments{
\item{atoms}{a list of literals}

\item{identifiers}{a list of identifiers}

\item{access_options}{}
}
\value{
\code{ResourceDescriptionFramework}
}
\description{
Taxonomic Key Section Constructor
}
