% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxonomicNameUsage.R
\name{TaxonomicNameUsage}
\alias{TaxonomicNameUsage}
\title{Construct TaxonomicNameUsage}
\usage{
TaxonomicNameUsage(node)
}
\arguments{
\item{node}{XML node, contains the taxonomic name usage}
}
\value{
TaxonomicNameUsaage
}
\description{
TODO: should be parameterized with the XML schema (i.e. TaxOub) that the
constructor is using.
}
