% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{title}
\alias{title}
\title{Title Constructor}
\usage{
title(atoms, identifiers, access_options, schema_name, mongo_key)
}
\arguments{
\item{atoms}{a list of literals}

\item{identifiers}{a list of identifiers}

\item{access_options}{}
}
\value{
\code{ResourceDescriptionFramework}
}
\description{
Title Constructor
}
