% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{treatment}
\alias{treatment}
\title{Treatment Constructor}
\usage{
treatment(atoms, identifiers, access_options, schema_name, mongo_key)
}
\arguments{
\item{atoms}{a list of literals}

\item{identifiers}{a list of identifiers}

\item{access_options}{}
}
\value{
\code{ResourceDescriptionFramework}
}
\description{
Treatment Constructor
}
