% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic-elements.R
\docType{data}
\name{Treatment}
\alias{Treatment}
\title{Treatment Section}
\format{An object of class \code{identifier} of length 4.}
\usage{
Treatment
}
\description{
Treatment Section
}
\keyword{datasets}
