% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nactem.R
\name{underscore_taxonomic_names}
\alias{underscore_taxonomic_names}
\title{Underscore Taxonimic Names in an XML Document Object}
\usage{
underscore_taxonomic_names(xmldoc, TNU_markup)
}
\arguments{
\item{xmldoc}{an XML2 object in which to underscore the names}

\item{TNU_markup}{a string with the markup of taxonomic names/ taxonomic name usages}
}
\value{
a character vector of underscored names (only binomial part)
}
\description{
Underscores taxonimic names (only keeps the binomial part).
}
\details{
Side-effect: modifies the XML object.
}
