% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nactem.R
\name{xml2nactem}
\alias{xml2nactem}
\title{Convert an XML File to Plain Text NaCTeM Format}
\usage{
xml2nactem(xml_filename, TNU_markup, PTNU_markup, regPTNU_markup, xmlschema)
}
\arguments{
\item{xml_filename}{the filename (one XML file) to convert to NaCTeM plain text}

\item{xmlschema}{the schema of the XML document; will be assigned as class to the XML object, once the file is read}
}
\value{
a list of the names that have been extracted and the plain text or an error
}
\description{
Convert an XML File to Plain Text NaCTeM Format
}
