% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{xml2rdf}
\alias{xml2rdf}
\title{RDF-ization of a Single XML File}
\usage{
xml2rdf(filename, xml_schema = taxonx, access_options, serialization_dir,
  reporcess = c("root_extractor"))
}
\arguments{
\item{filename}{locator, e.g. file path or URL, of the XML resource}

\item{xml_schema}{XML schema of the XML resource; one of
taxpub taxonx or plazi_int}

\item{access_options}{endpoint to database to which to submit}

\item{serialization_dir}{directory where to dump intermediate serializations}

\item{serialization_format}{default is "turtle"}

\item{reprocess}{\code{logical}}
}
\value{
\code{logical}. TRUE if everything went OK. FALSE if there was a problem
}
\description{
Converts an XML file to RDF and submits it to a triple store. Writes
the serialization in a file in the file system.
}
