% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\docType{data}
\name{XmlSchema}
\alias{XmlSchema}
\title{XML Node Schema}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
XmlSchema
}
\description{
Defines the XML schema of a node.
}
\section{Fields}{

\describe{
\item{\code{schema_name}}{the name of the schema}

\item{\code{file_pattern}}{regular expression pattern matching the file-name extension}

\item{\code{extension}}{the file-name extension}

\item{\code{prefix}}{the prefix that these documents have in the URI}

\item{\code{atoms}}{named character vector of xpath locations}

\item{\code{constructor}}{an RDF constructor function that can be called on
a list of atoms extractor from a node in the schema}

\item{\code{components}}{a list of \code{XmlSchema} object containing the nested
components in the node}
}}

\examples{
taxonx = XmlSchema$new(schema_name = "taxonx", file_pattern = ".*\\\\.xml")

}
\keyword{datasets}
